+++ app/Http/Controllers/Api/V1/Auth/PasswordResetController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespacesuperadmin App\Http\Controllers\Api\V1\Auth;
+namespace App\Http\Controllers\Api\V1\Auth;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Auth\Events\PasswordReset;